requirements.txt
chromadb>=0.4.0
langchain>=0.1.0
ollama>=0.1.0
pandas>=2.0.0
pyodbc>=5.0.0
sqlalchemy>=2.0.0
streamlit>=1.30.0
sentence-transformers>=2.2.0
pyyaml>=6.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rdata>=0.10.0
